# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vulnerability Assistance Platform API",
    "version": "1.0.0",
    "description": "API para plataforma de acompanhamento de pessoas em vulnerabilidade urbana",
    "contact": {
      "name": "API Support",
      "email": "support@vulnerability-platform.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT válido por 5 dias"
      }
    },
    "schemas": {
...
      "Checkpoint": {
        "type": "object",
        "required": [
          "project_id",
          "nome",
          "peso"
        ],
        "properties": {
          "checkpoint_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID único do checkpoint (auto-gerado)",
            "readOnly": true
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do projeto"
          },
          "nome": {
            "type": "string",
            "maxLength": 255,
            "description": "Nome do checkpoint"
          },
          "peso": {
            "type": "integer",
            "minimum": 0,
            "description": "Peso do checkpoint"
          },
          "data_criacao": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação"
          }
        }
      },
      "CheckpointProgress": {
        "type": "object",
        "required": [
          "person_id",
          "checkpoint_id"
        ],
        "properties": {
          "progress_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID único do progresso (auto-gerado)",
            "readOnly": true
          },
          "person_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID da pessoa"
          },
          "checkpoint_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do checkpoint"
          },
          "status": {
            "type": "string",
            "enum": [
              "pendente",
              "em_progresso",
              "concluido"
            ],
            "description": "Status do progresso"
          },
          "data_atualizacao": {
            "type": "string",
            "format": "date-time",
            "description": "Data da última atualização"
          },
          "observacoes": {
            "type": "string",
            "nullable": true,
            "description": "Observações sobre o progresso"
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica se a operação foi bem-sucedida"
          },
          "data": {
            "type": "object",
            "description": "Dados retornados"
          },
          "message": {
            "type": "string",
            "description": "Mensagem de resposta"
          },
          "count": {
            "type": "integer",
            "description": "Número de itens (para listas)"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Mensagem de erro"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email do usuário"
          },
          "senha": {
            "type": "string",
            "description": "Senha do usuário"
          }
        },
        "example": {
          "email": "joao@example.com",
          "senha": "senha123"
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string",
                "description": "Token JWT válido por 5 dias"
              },
              "user": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "nome": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "tipo": {
                    "type": "string",
                    "enum": [
                      "MISSIONARIO",
                      "INVESTIDOR"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ATIVO",
                      "INATIVO"
                    ]
                  }
                }
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Realizar login e obter token JWT",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/api/auth/validate": {
      "get": {
        "summary": "Validar token JWT",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token válido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido ou expirado"
          }
        }
      }
    },
    "/api/checkpoint-progress": {
      "post": {
        "summary": "Registrar progresso em checkpoint",
        "tags": [
          "Checkpoint Progress"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointProgress"
              },
              "example": {
                "person_id": "uuid-da-pessoa",
                "checkpoint_id": "uuid-do-checkpoint",
                "status": "em_progresso",
                "observacoes": "Primeira consulta realizada com sucesso"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Progresso registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar progressos",
        "tags": [
          "Checkpoint Progress"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "person_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtrar por ID da pessoa"
          },
          {
            "in": "query",
            "name": "checkpoint_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtrar por ID do checkpoint"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pendente",
                "em_progresso",
                "concluido"
              ]
            },
            "description": "Filtrar por status"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de progressos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkpoint-progress/{id}": {
      "get": {
        "summary": "Buscar progresso por ID",
        "tags": [
          "Checkpoint Progress"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do progresso"
          }
        ],
        "responses": {
          "200": {
            "description": "Progresso encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Progresso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar progresso",
        "tags": [
          "Checkpoint Progress"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do progresso"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointProgress"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Progresso atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remover progresso",
        "tags": [
          "Checkpoint Progress"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do progresso"
          }
        ],
        "responses": {
          "200": {
            "description": "Progresso removido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Progresso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkpoints": {
      "post": {
        "summary": "Criar novo checkpoint",
        "tags": [
          "Checkpoints"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Checkpoint"
              },
              "example": {
                "project_id": "uuid-do-projeto",
                "nome": "Primeira Consulta",
                "peso": 10
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Checkpoint criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar checkpoints",
        "tags": [
          "Checkpoints"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtrar por ID do projeto"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de checkpoints",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkpoints/{id}": {
      "get": {
        "summary": "Buscar checkpoint por ID",
        "tags": [
          "Checkpoints"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do checkpoint"
          }
        ],
        "responses": {
          "200": {
            "description": "Checkpoint encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Checkpoint não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar checkpoint",
        "tags": [
          "Checkpoints"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do checkpoint"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Checkpoint"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkpoint atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deletar checkpoint",
        "tags": [
          "Checkpoints"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID do checkpoint"
          }
        ],
        "responses": {
          "200": {
            "description": "Checkpoint deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Checkpoint não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}